requirements.txt
fastapi>=0.114.2
uvicorn[standard]>=0.30.6
SQLAlchemy>=2.0.35
alembic>=1.13.3
pydantic>=2.9.2
psycopg[binary]>=3.2.3
python-dotenv>=1.0.1

email-validator>=2.2.0

psycopg2-binary>=2.9.9
